chore(data): add 26th single tite and move title type
