+++ src/server/actors/Discography/models.ts
@@ -2,12 +2,9 @@ import { Site } from 'server/types/commons';
 import { SongTitle } from 'server/actors/Songs/constants/songTitle';
 import { SongType } from 'server/actors/Songs/constants/songType';
 import { MemberNameKey } from 'server/actors/Members/constants/memberName';
-import {
-  CdTitle,
-  CdType,
-  DiscographyType,
-} from 'server/actors/Discography/types';
+import { CdType, DiscographyType } from 'server/actors/Discography/types';
 import { UnitName } from 'server/actors/Units/constants/unitName';
+import { CdTitle } from 'server/actors/Discography/constants/cdTitle';
 
 export type CdSongRaw = {
   number: number;
